requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosqlite==0.19.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
frozenlist==1.6.0
h11==0.14.0
httpcore==1.0.8
httpx==0.25.2
idna==3.10
magic-filter==1.0.12
multidict==6.4.3
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.0.0
python-telegram-bot==20.7
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
yarl==1.20.0
